# jxl-modular: Try various optimizations