<details>
<summary>Benchmark result (x86_64)</summary>

```
lumine-paimon.d0-e7/preferred-color
                        time:   [102.55 ms 103.09 ms 103.67 ms]
                        thrpt:  [17.127 Melem/s 17.224 Melem/s 17.315 Melem/s]
                 change:
                        time:   [-2.3218% -1.5582% -0.7626%] (p = 0.00 < 0.05)
                        thrpt:  [+0.7684% +1.5829% +2.3770%]
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

minecraft.d0-e6/preferred-color
                        time:   [89.826 ms 90.212 ms 90.599 ms]
                        thrpt:  [40.689 Melem/s 40.864 Melem/s 41.039 Melem/s]
                 change:
                        time:   [-7.9255% -7.3408% -6.7342%] (p = 0.00 < 0.05)
                        thrpt:  [+7.2205% +7.9224% +8.6077%]
                        Performance has improved.

srgb.d0-e1/preferred-color
                        time:   [12.066 ms 12.103 ms 12.144 ms]
                        thrpt:  [118.28 Melem/s 118.68 Melem/s 119.04 Melem/s]
                 change:
                        time:   [-6.1269% -5.7340% -5.3368%] (p = 0.00 < 0.05)
                        thrpt:  [+5.6377% +6.0828% +6.5268%]
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

nahida-motion.d1-e7/preferred-color
                        time:   [57.398 ms 57.479 ms 57.560 ms]
                        thrpt:  [64.044 Melem/s 64.135 Melem/s 64.225 Melem/s]
                 change:
                        time:   [-3.0342% -2.8442% -2.6579%] (p = 0.00 < 0.05)
                        thrpt:  [+2.7304% +2.9275% +3.1292%]
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild

starrail.d1-e6/preferred-color
                        time:   [57.377 ms 57.474 ms 57.574 ms]
                        thrpt:  [64.029 Melem/s 64.141 Melem/s 64.249 Melem/s]
                 change:
                        time:   [-2.7467% -2.5262% -2.3053%] (p = 0.00 < 0.05)
                        thrpt:  [+2.3597% +2.5917% +2.8243%]
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

genshin-cafe.d2-e6-epf2/preferred-color
                        time:   [94.150 ms 94.309 ms 94.471 ms]
                        thrpt:  [71.641 Melem/s 71.764 Melem/s 71.885 Melem/s]
                 change:
                        time:   [-4.7657% -4.3981% -4.0614%] (p = 0.00 < 0.05)
                        thrpt:  [+4.2333% +4.6004% +5.0042%]
                        Performance has improved.

genshin-cafe.d2-e6-epf3/preferred-color
                        time:   [114.02 ms 114.28 ms 114.62 ms]
                        thrpt:  [59.046 Melem/s 59.222 Melem/s 59.359 Melem/s]
                 change:
                        time:   [+4.1649% +4.4929% +4.8871%] (p = 0.00 < 0.05)
                        thrpt:  [-4.6594% -4.2998% -3.9984%]
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
```

Result of `genshin-cafe.d2-e6-epf3` seems a bit off...
</details>